Move MatchaState mock object to conftest
